+++ static/html/signup.html
@@ -3,89 +3,21 @@
     <head>
         <meta charset="UTF-8" />
         <meta name="viewport" content="width=device-width, initial-scale=1.0" />
+        <link rel="icon" type="image/svg+xml" sizes="32x32" href="../svg/logo_favicon.svg"/>
         <link rel="stylesheet" href="../css/style.css" type="text/css" />
-        <script src="../js/handlebars.js"></script>
-        <script type="module">
-            "use strict";
-
-            import { Image } from "../js/components/backgroundImage/BackgroundImage.js";
-
-            const root = document.querySelector(".page");
-
-            const signupConfig = {
-                images: {
-                    left: {
-                        side: "left",
-                        picture: "undraw_performance_overview",
-                    },
-                    right: {
-                        side: "right",
-                        picture: "undraw_team_collaboration",
-                    },
-                },
-            };
-
-            const images = new Image(root, config.images);
-            images.render();
-        </script>
+        <script src="../js/handlebars.runtime.js"></script>
+        <script src="../js/components/backgroundImage/backgroundImage.precompile.js"></script>
+        <script src="../js/components/signLocation/signLocation.precompile.js"></script>
+        <script src="../js/components/mainLogo/mainLogo.precompile.js"></script>
+        <script src="../js/components/formTitle/formTitle.precompile.js"></script>
+        <script src="../js/components/formComponent/formComponent.precompile.js"></script>
+        <script src="../js/components/formInput/formInput.precompile.js"></script>
+        <script src="../js/components/signHref/signHref.precompile.js"></script>
+        <script src="../js/components/signButton/signButton.precompile.js"></script>
+        <script type="module" src="../js/index.js"></script>
         <title>Tabula: Signup</title>
     </head>
     <body>
-        <div class="page">
-            <!-- <img
-                class="page__image page__image_right"
-                src="../svg/undraw_team_collaboration.svg"
-            />
-            <img
-                class="page__image page__image_left"
-                src="../svg/undraw_performance_overview.svg"
-            /> -->
-            <div class="sign-location">
-                <h1 class="sign-location__header">
-                    <img src="../svg/logo.svg" />
-                    <span class="sign-location__name"></span>Tabula
-                </h1>
-                <h2 class="sign-location__header-title">Регистрация</h2>
-                <form
-                    class="sign-form sign-form_centered"
-                    action="/"
-                    method="post"
-                    enctype="application/x-www-form-urlencoded"
-                >
-                    <div class="sign-form__container">
-                        <img src="../svg/person.svg" />
-                        <input
-                            type="text"
-                            class="sign-form__input"
-                            placeholder="Email"
-                        />
-                    </div>
-                    <div class="sign-form__container">
-                        <img src="../svg/lock.svg" />
-                        <input
-                            type="password"
-                            class="sign-form__input"
-                            placeholder="Пароль"
-                        />
-                    </div>
-                    <div class="sign-form__container">
-                        <img src="../svg/key.svg" />
-                        <input
-                            type="password"
-                            class="sign-form__input"
-                            placeholder="Повторите пароль"
-                        />
-                    </div>
-                    <a class="href-sign" href="signin.html"
-                        >Уже есть аккаунт?</a
-                    >
-                    <input
-                        type="submit"
-                        class="button-sign"
-                        value="Зарегистрироваться"
-                    />
-                </form>
-            </div>
-        </div>
+        <div class="page"></div>
     </body>
 </html>
