--- static/js/components/backgroundImage/BackgroundImage.hbs
@@ -1,4 +1,0 @@
-<img
-        class="page__image page__image_{{this.side}}"
-        src="{{this.picture}}.svg"
-      />
