+++ static/js/components/backgroundImage/backgroundImage.hbs
@@ -1,0 +1,1 @@
+<img class="page__image page__image_{{this.side}}" src="{{this.picture}}.svg" />
