+++ static/js/components/formComponent/formComponent.hbs
@@ -1,0 +1,1 @@
+<form class="sign-form sign-form_centered" action="/" method="post" enctype="application/x-www-form-urlencoded"></form>
