+++ static/js/components/formComponent/formComponent.precompile.js
@@ -1,0 +1,6 @@
+(function() {
+  var template = Handlebars.template, templates = Handlebars.templates = Handlebars.templates || {};
+templates['formComponent.hbs'] = template({"compiler":[8,">= 4.3.0"],"main":function(container,depth0,helpers,partials,data) {
+    return "<form class=\"sign-form sign-form_centered\" action=\"/\" method=\"post\" enctype=\"application/x-www-form-urlencoded\"></form>";
+},"useData":true});
+})();
