+++ static/js/components/formInput/formInput.hbs
@@ -1,0 +1,8 @@
+<div class="sign-form__container">
+    <img src="{{this.icon}}.svg" />
+    <input
+        type="{{this.type}}"
+        class="sign-form__input"
+        placeholder="{{this.placeholder}}"
+    />
+</div>
