+++ static/js/components/formTitle/formTitle.hbs
@@ -1,0 +1,1 @@
+<h2 class="sign-location__header-title">{{this.title}}</h2>
