+++ static/js/components/mainLogo/mainLogo.hbs
@@ -1,0 +1,4 @@
+<h1 class="sign-location__header">
+<img src="{{this.logo}}.svg" />
+<span class="sign-location__name"></span>{{this.title}}
+ </h1>
