+++ static/js/components/mainLogo/mainLogo.precompile.js
@@ -1,0 +1,17 @@
+(function() {
+  var template = Handlebars.template, templates = Handlebars.templates = Handlebars.templates || {};
+templates['mainLogo.hbs'] = template({"compiler":[8,">= 4.3.0"],"main":function(container,depth0,helpers,partials,data) {
+    var alias1=container.lambda, alias2=container.escapeExpression, lookupProperty = container.lookupProperty || function(parent, propertyName) {
+        if (Object.prototype.hasOwnProperty.call(parent, propertyName)) {
+          return parent[propertyName];
+        }
+        return undefined
+    };
+
+  return "<h1 class=\"sign-location__header\">\n<img src=\""
+    + alias2(alias1((depth0 != null ? lookupProperty(depth0,"logo") : depth0), depth0))
+    + ".svg\" />\n<span class=\"sign-location__name\"></span>"
+    + alias2(alias1((depth0 != null ? lookupProperty(depth0,"title") : depth0), depth0))
+    + "\n </h1>";
+},"useData":true});
+})();
