+++ static/js/components/precompiler
@@ -1,0 +1,13 @@
+#!/bin/bash
+if [ -f /JSscripts.txt ]; then
+    touch JSscripts.txt
+    echo > 1
+else
+    echo -n > JSscripts.txt
+fi
+ls -d */ | rev | cut -c 2- | rev | while read el; do
+    cd ./$el
+    handlebars $el.hbs -f $el.precompile.js
+    cd ..
+    echo "<script src="../js/components/$el/$el.precompile.js"></script>" >> ./JSscripts.txt
+done
