+++ static/js/components/signButton/signButton.hbs
@@ -1,0 +1,1 @@
+<input type="submit" class="button-sign" value="{{this.text}}" />
