+++ static/js/components/signButton/signButton.js
@@ -1,0 +1,18 @@
+export class SignButton {
+    #parent;
+    #config;
+
+    constructor(parent, config) {
+        this.#parent = parent;
+        this.#config = config;
+    }
+
+    get items() {
+        return { text: this.#config.text };
+    }
+
+    render() {
+        const template = Handlebars.templates["signButton.hbs"];
+        this.#parent.innerHTML += template(this.items);
+    }
+}
