+++ static/js/components/signButton/signButton.precompile.js
@@ -1,0 +1,6 @@
+(function() {
+  var template = Handlebars.template, templates = Handlebars.templates = Handlebars.templates || {};
+templates['signButton.hbs'] = template({"compiler":[8,">= 4.3.0"],"main":function(container,depth0,helpers,partials,data) {
+    return "<input\n                        type=\"submit\"\n                        class=\"button-sign\"\n                        value=\"Зарегистрироваться\"\n                    />";
+},"useData":true});
+})();
