+++ static/js/components/signHref/signHref.hbs
@@ -1,0 +1,1 @@
+<a class="href-sign" href="{{this.url}}.html">{{this.text}}</a>
