+++ static/js/components/signLocation/signLocation.hbs
@@ -1,0 +1,1 @@
+<div class="sign-location"></div>
