+++ static/js/components/signLocation/signLocation.js
@@ -1,0 +1,13 @@
+export class Location {
+    #parent;
+
+    constructor(parent) {
+        this.#parent = parent;
+    }
+
+    render() {
+        const template = Handlebars.templates["signLocation.hbs"];
+
+        this.#parent.innerHTML += template({});
+    }
+}
