+++ static/js/components/signLocation/signLocation.precompile.js
@@ -1,0 +1,6 @@
+(function() {
+  var template = Handlebars.template, templates = Handlebars.templates = Handlebars.templates || {};
+templates['signLocation.hbs'] = template({"compiler":[8,">= 4.3.0"],"main":function(container,depth0,helpers,partials,data) {
+    return "<div class=\"sign-location\"></div>\n";
+},"useData":true});
+})();
