--- static/js/formInput/FormInput.hbs
@@ -1,8 +1,0 @@
-<div class="sign-form__container">
-    <img src="../svg/{{svgIcon}}" />
-    <input
-        type="{{type}}"
-        class="sign-form__input"
-        placeholder="{{placeholder}}"
-    />
-</div>
